requirements/base.txt
# Python --version 3.10.6

celery==5.2.7
Django==3.2.16
django-baton==2.5.1
django-celery-beat==2.4.0
django-environ==0.9.0
djangorestframework==3.14.0
redis==4.3.4
requests==2.28.1
django-allauth==0.51.0
django-auth-style==0.8.1
